Move numbered rule renaming visitor to record/parsers
